refactor: ActiveBids section
